+++ sql_models/model_schematisation.py
@@ -54,7 +54,7 @@ class CrossSectionDefinition(Base):
     width = Column(Float)
     height = Column(Float)
     shape = Column(Integer)  # PROFILE_SHAPES
-    _code = Column(String(100))  # nullable=False
+    code = Column(String(100), default='', nullable=False)
 
 
 class ConnectionNode(Base):
@@ -76,8 +76,7 @@ class ConnectionNode(Base):
                                  nullable=True)
 
     # extra fields:
-    _code = Column(String(100), nullable=True)
-    _basin_code = Column(String(4), nullable=True)
+    code = Column(String(100), default='', nullable=True)
 
     manhole = relationship("Manhole",
                            uselist=False,
@@ -115,7 +114,7 @@ class Manhole(Base):
     connection_node_id = Column(
             Integer, ForeignKey(ConnectionNode.__tablename__ + ".id"),
             nullable=False,
-            unique=True)  # extra compared to original
+            unique=True)
     connection_node = relationship(ConnectionNode,
                                    back_populates="manhole")
 
@@ -182,6 +181,7 @@ class CrossSectionLocation(Base):
     FRICTION_TYPE = Constants.FRICTION_TYPES
 
     id = Column(Integer, primary_key=True)
+    code = Column(String(100), default='', nullable=False)
     channel_id = Column(
         Integer, ForeignKey("v2_channel.id"),
         nullable=False)
@@ -246,8 +246,6 @@ class Pipe(Base):
 
     material = Column(Integer)  # MATERIALS
     pipe_quality = Column(Float)
-
-    _basin = Column(String(4), nullable=True)
 
 
 class Culvert(Base):
@@ -411,6 +409,8 @@ class Pumpstation(Base):
 
     sewerage = Column(Boolean, default=False)
     classification = Column(Integer)  # in use?
+    type_ = Column(Integer, nullable=True, default=1, name='type')
+
 
     # relation ships
     connection_node_start_id = Column(
@@ -428,18 +428,17 @@ class Pumpstation(Base):
             ConnectionNode, foreign_keys=connection_node_end_id)
 
     # pump details
-    start_level_delivery_side = Column(Float)
-    stop_level_delivery_side = Column(Float)
-    start_level_suction_side = Column(Float)
-    stop_level_suction_side = Column(Float)
+    start_level = Column(Float)
+    lower_stop_level = Column(Float)
+    upper_stop_level = Column(Float)
     capacity = Column(Float)
 
 
 class Obstacle(Base):
-    __tablename__ = 'v2_levee'
+    __tablename__ = 'v2_obstacle'
 
     id = Column(Integer, primary_key=True)
-    _code = Column(String(100), nullable=True)  # extra
+    code = Column(String(100), default='', nullable=False)
 
     crest_level = Column(Float)
     the_geom = Column(Geometry(geometry_type='LINESTRING',
@@ -449,12 +448,12 @@ class Obstacle(Base):
 
 
 class Levee(Base):
-    __tablename__ = 'v2_obstacle'
+    __tablename__ = 'v2_levee'
 
     LEVEE_MATERIALS = Constants.LEVEE_MATERIALS
 
     id = Column(Integer, primary_key=True)
-    _code = Column(String(100), nullable=True)  # extra
+    code = Column(String(100), default='', nullable=False)
 
     crest_level = Column(Float)
     the_geom = Column(Geometry(geometry_type='LINESTRING',